Changed feedBoundaryIntoFsgrid so that it doesn't calculate coupling every timestep. Also fixed some issues so it compiles. Also moved computeCoupling implementation from gridglue.cpp to gridglue.hpp, as it is a template function and cannot be used elsewhere otherwise.
