+++ fieldsolver/gridGlue.hpp
@@ -32,6 +32,13 @@ enum FieldsToCommunicate {
    N_FIELDSTOCOMMUNICATE
 };
 
+
+void filterMoments(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
+                           FsGrid< std::array<Real, fsgrids::moments::N_MOMENTS>, FS_STENCIL_WIDTH> & momentsGrid,
+                           FsGrid< fsgrids::technical, FS_STENCIL_WIDTH> & technicalGrid);
+
+
+
 std::vector<CellID> mapDccrgIdToFsGridGlobalID(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
 					       CellID dccrgID);
 
@@ -102,4 +109,71 @@ int getNumberOfCellsOnMaxRefLvl(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geomet
 
 void feedBoundaryIntoFsGrid(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
 			const std::vector<CellID>& cells,
-			FsGrid< fsgrids::technical, FS_STENCIL_WIDTH> & technicalGrid);
+			FsGrid< fsgrids::technical, FS_STENCIL_WIDTH> & technicalGrid,
+         std::map<int, std::set<CellID> >& onDccrgMapRemoteProcess,
+         std::map<int, std::set<CellID> >& onFsgridMapRemoteProcess,
+         std::map<CellID, std::vector<int64_t> >&  onFsgridMapCells);
+
+/*Compute coupling DCCRG <=> FSGRID 
+
+  onDccrgMapRemoteProcess   maps fsgrid processes (key) => set of dccrg cellIDs owned by current rank that map to  the fsgrid cells owned by fsgrid process (val)
+
+  onFsgridMapRemoteProcess  maps dccrg processes  (key) => set of dccrg cellIDs owned by dccrg-process that map to current rank fsgrid cells 
+  onFsgridMapCells          maps remote dccrg CellIDs to local fsgrid cells
+*/
+
+// this function is declared here as it is a template function
+
+template <typename T, int stencil> void computeCoupling(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
+							const std::vector<CellID>& cells,
+							FsGrid< T, stencil>& momentsGrid,
+							std::map<int, std::set<CellID> >& onDccrgMapRemoteProcess,
+							std::map<int, std::set<CellID> >& onFsgridMapRemoteProcess,
+							std::map<CellID, std::vector<int64_t> >& onFsgridMapCells
+							) {
+    
+  //sorted list of dccrg cells. cells is typicall already sorted, but just to make sure....
+  std::vector<CellID> dccrgCells = cells;
+  std::sort(dccrgCells.begin(), dccrgCells.end());
+
+  //make sure the datastructures are clean
+  onDccrgMapRemoteProcess.clear();
+  onFsgridMapRemoteProcess.clear();
+  onFsgridMapCells.clear();
+  
+  
+  //size of fsgrid local part
+  const std::array<FsGridTools::FsIndex_t, 3> gridDims(momentsGrid.getLocalSize());
+  
+ 
+  //Compute what we will receive, and where it should be stored
+  for (FsGridTools::FsIndex_t k=0; k<gridDims[2]; k++) {
+    for (FsGridTools::FsIndex_t j=0; j<gridDims[1]; j++) {
+      for (FsGridTools::FsIndex_t i=0; i<gridDims[0]; i++) {
+        const std::array<FsGridTools::FsIndex_t, 3> globalIndices = momentsGrid.getGlobalIndices(i,j,k);
+        const dccrg::Types<3>::indices_t  indices = {{(uint64_t)globalIndices[0],
+                        (uint64_t)globalIndices[1],
+                        (uint64_t)globalIndices[2]}}; //cast to avoid warnings
+        CellID dccrgCell = mpiGrid.get_existing_cell(indices, 0, mpiGrid.mapping.get_maximum_refinement_level());
+        
+        int process = mpiGrid.get_process(dccrgCell);
+        FsGridTools::LocalID fsgridLid = momentsGrid.LocalIDForCoords(i,j,k);
+        //int64_t  fsgridGid = momentsGrid.GlobalIDForCoords(i,j,k);
+        onFsgridMapRemoteProcess[process].insert(dccrgCell); //cells are ordered (sorted) in set
+        onFsgridMapCells[dccrgCell].push_back(fsgridLid);
+      }
+    }
+  }
+
+  // Compute where to send data and what to send
+  for(uint64_t i=0; i< dccrgCells.size(); i++) {
+     //compute to which processes this cell maps
+     std::vector<CellID> fsCells = mapDccrgIdToFsGridGlobalID(mpiGrid, dccrgCells[i]);
+
+     //loop over fsgrid cells which this dccrg cell maps to
+     for (auto const &fsCellID : fsCells) {
+       int process = momentsGrid.getTaskForGlobalID(fsCellID).first; //process on fsgrid
+       onDccrgMapRemoteProcess[process].insert(dccrgCells[i]); //add to map
+     }    
+  }
+}
